⚡️ Render docs codeblocks on server
Rather than in UI, to avoid layout changes
