+++ packages/docs/src/lib/components/CodeBlock.svelte
@@ -1,24 +1,12 @@
 <script lang="ts">
-  import { getSingletonHighlighter, type BundledLanguage } from "shiki";
-
   import CopyButton from "./CopyButton.svelte";
   import "./CodeBlock.css";
 
-  let { code, lang }: { code: string; lang?: BundledLanguage } = $props();
-
-  const highlighter = async (highCode: string, highLang?: BundledLanguage) => {
-    const langToLoad = highLang || "text";
-    const highlighterTool = await getSingletonHighlighter({
-      themes: ["nord"],
-      langs: [langToLoad],
-    });
-    await highlighterTool.loadTheme("nord");
-    const html = highlighterTool.codeToHtml(highCode, {
-      lang: langToLoad,
-      theme: "nord",
-    });
-    return html;
-  };
+  let {
+    code,
+    codeHtml,
+    lang,
+  }: { code: string; codeHtml: string; lang?: string } = $props();
 </script>
 
 <div class="code-block-wrapper rounded-md bg-codeblock-500">
@@ -32,8 +20,6 @@
     {/if}
     <CopyButton textToCopy={code} />
   </div>
-  {#await highlighter(code, lang) then highlightedCode}
-    <!-- eslint-disable-next-line svelte/no-at-html-tags -->
-    {@html highlightedCode}
-  {/await}
+  <!-- eslint-disable-next-line svelte/no-at-html-tags -->
+  {@html codeHtml}
 </div>
